dev: updated testcase for wrapper
